requirements.txt
google-generativeai>=0.8.0
Pillow>=10.0.0
opencv-python>=4.9.0
pytesseract>=0.3.10
numpy>=1.26.0